+++ examples/vue-apollo/src/pages/sign-in/EmailPasswordless.vue
@@ -1,6 +1,6 @@
 <template>
   <email-passwordless />
   <v-btn class="my-1" block variant="text" color="primary" to="/signin">
-    &#8592; Other Login Options
+    &#8592; Other Sign-in Options
   </v-btn>
 </template>