+++ packages/core/src/machines/change-password.ts
@@ -3,6 +3,7 @@ import { assign, createMachine, send } from 'xstate'
 import { AuthClient } from '../client'
 import { ErrorPayload, INVALID_PASSWORD_ERROR } from '../errors'
 import { nhostApiClient } from '../hasura-auth'
+import { ChangePasswordResponse } from '../types'
 import { isValidPassword } from '../validators'
 
 export type ChangePasswordContext = {
@@ -12,9 +13,14 @@ export type ChangePasswordEvents =
   | {
       type: 'REQUEST'
       password?: string
+      ticket?: string
     }
   | { type: 'SUCCESS' }
   | { type: 'ERROR'; error: ErrorPayload | null }
+
+export type ChangePasswordServices = {
+  requestChange: { data: ChangePasswordResponse }
+}
 
 export type ChangePasswordMachine = ReturnType<typeof createChangePasswordMachine>
 
@@ -24,7 +30,8 @@ export const createChangePasswordMachine = ({ backendUrl, interpreter }: AuthCli
     {
       schema: {
         context: {} as ChangePasswordContext,
-        events: {} as ChangePasswordEvents
+        events: {} as ChangePasswordEvents,
+        services: {} as ChangePasswordServices
       },
       tsTypes: {} as import('./change-password.typegen').Typegen0,
       preserveActionOrder: true,
@@ -66,6 +73,7 @@ export const createChangePasswordMachine = ({ backendUrl, interpreter }: AuthCli
       actions: {
         saveInvalidPasswordError: assign({ error: (_) => INVALID_PASSWORD_ERROR }),
         saveRequestError: assign({
+          // * Untyped action payload. See https://github.com/statelyai/xstate/issues/3037
           error: (_, { data: { error } }: any) => error
         }),
         reportError: send((ctx) => ({ type: 'ERROR', error: ctx.error })),
@@ -75,10 +83,10 @@ export const createChangePasswordMachine = ({ backendUrl, interpreter }: AuthCli
         invalidPassword: (_, { password }) => !isValidPassword(password)
       },
       services: {
-        requestChange: (_, { password }) =>
-          api.post<string, { data: { error?: ErrorPayload } }>(
+        requestChange: (_, { password, ticket }) =>
+          api.post<string, ChangePasswordResponse>(
             '/user/password',
-            { newPassword: password },
+            { newPassword: password, ticket: ticket },
             {
               headers: {
                 authorization: `Bearer ${interpreter?.state.context.accessToken.value}`