+++ packages/hasura-storage-js/tests/helpers/handlers/index.ts
@@ -1,2 +1,0 @@
-export * from './uploadFileHandlers'
-export * from './authHandlers'