+++ templates/functions/node/typescript.ts
@@ -1,7 +1,0 @@
-import { Request, Response } from "express";
-
-export default (req: Request, res: Response) => {
-  res
-    .status(200)
-    .send(`Nhost, from Typescript, pays it's respects to ${req.query.name}!`);
-};