+++ templates/web/nextjs-apollo/utils/nhost.ts
@@ -1,7 +1,0 @@
-import { NhostClient } from "@nhost/nhost-js";
-
-const nhost = new NhostClient({
-  backendUrl: "http://localhost:1337",
-});
-
-export { nhost };